# docs/commons/openapi.yaml
openapi: 3.0.3
info:
  title: Reals Backend API
  version: 1.0.0
  description: Current backend contract for the Reals mobile client.
servers:
  - url: http://10.0.2.2:8080
    description: Android emulator against local backend
  - url: http://localhost:8080
    description: Local backend from the same machine
security:
  - firebaseAppCheck: []
    bearerAuth: []
...
paths:
...
  /api/matches/{matchId}:
    get:
      tags: [Matches]
      summary: Get match details
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          $ref: '#/components/responses/Match'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/matches/{matchId}/chat:
    get:
      tags: [Matches]
      summary: Get active first chat for match
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Active first chat with partner and participant decisions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirstChatResponse'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
  /api/matches/{matchId}/visual-profile:
    get:
      tags: [Matches]
      summary: Get partner profile visible during visual phase or later
      description: Returns the partner profile and visual-review metadata from the authenticated user's perspective, including whether the current user already submitted their visual personal message.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Partner visual profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisualProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/matches/{matchId}/chat-decision:
    post:
      tags: [Matches]
      summary: Submit first-chat continuation decision
      description: APPROVED requires current legal requirements to be satisfied. REJECTED remains available without legal compliance.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        $ref: '#/components/requestBodies/ChatDecision'
      responses:
        '200':
          $ref: '#/components/responses/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatDecisionConflict'
  /api/matches/{matchId}/visual-decision:
    post:
      tags: [Matches]
      summary: Submit visual decision
      description: APPROVED requires current legal requirements to be satisfied. REJECTED remains available without legal compliance. Partner personal messages are optional and may remain unread before any visual decision.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        $ref: '#/components/requestBodies/VisualDecision'
      responses:
        '200':
          $ref: '#/components/responses/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/matches/{matchId}/personal-messages/me:
    put:
      tags: [Matches]
      summary: Store authenticated user's personal visual-review message
      description: Stores the authenticated user's visual personal message once. A second submission returns 409 Conflict and does not overwrite the original message.
      parameters:
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        $ref: '#/components/requestBodies/PersonalMessage'
      responses:
        '204':
          description: Message stored
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/matches/{matchId}/personal-messages/partner:
    get:
      tags: [Matches]
      summary: Get partner personal visual-review message
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Partner message; message is null if not submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerMessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/matches/{matchId}/block:
    post:
      summary: Durably block the other participant in a match
      description: Creates a directional MANUAL block and immediately contains active interactions for the pair. No request body is required.
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: A new directional block was created
        '200':
          description: The same directional block already existed
        '403':
          description: The authenticated user is not a match participant
        '409':
          description: Interaction conflict, including USER_PAIR_BLOCKED
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    firebaseAppCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
      description: Firebase App Check token for Android-facing API requests.
  parameters:
    MatchId:
      name: matchId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
...
    ChatDecision:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatDecisionRequest'
    VisualDecision:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VisualDecisionRequest'
    PersonalMessage:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PersonalMessageRequest'
...
  responses:
...
    Match:
      description: Match
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MatchResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Domain conflict or invalid state transition. Only operations protected by the legal gate may also return stable code `LEGAL_ACTION_REQUIRED` when current legal document actions are required before continuing; reuse of this response by another operation does not make that operation legal-gated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            legalActionRequired:
              value:
                code: LEGAL_ACTION_REQUIRED
                error: Conflict
                message: Current legal document actions are required before continuing
...
    ChatNotFound:
      description: "Chat not found. Stable code: `CHAT_NOT_FOUND`."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CHAT_NOT_FOUND
            error: Not Found
            message: Chat was not found
    ChatAvailabilityConflict:
      description: Chat is not currently readable or writable. Message sends may also return `CHAT_MUTUAL_CANCELLATION_PENDING`; active chat participation writes may also return `LEGAL_ACTION_REQUIRED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            legalActionRequired:
              value:
                code: LEGAL_ACTION_REQUIRED
                error: Conflict
                message: Current legal document actions are required before continuing
            chatNotAvailable:
              value:
                code: CHAT_NOT_AVAILABLE
                error: Conflict
                message: Chat is not available
            chatExpired:
              value:
                code: CHAT_EXPIRED
                error: Conflict
                message: Chat has expired
            mutualCancellationPending:
              value:
                code: CHAT_MUTUAL_CANCELLATION_PENDING
                error: Conflict
                message: A mutual cancellation request is pending
...
    ChatDecisionConflict:
      description: First-chat continuation decision is not currently available. APPROVED may also return `LEGAL_ACTION_REQUIRED`; REJECTED remains available without legal compliance.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            legalActionRequired:
              value:
                code: LEGAL_ACTION_REQUIRED
                error: Conflict
                message: Current legal document actions are required before continuing
            decisionNotAvailable:
              value:
                code: CHAT_DECISION_NOT_AVAILABLE
                error: Conflict
                message: Chat decision is not available
            decisionAlreadySubmitted:
              value:
                code: CHAT_DECISION_ALREADY_SUBMITTED
                error: Conflict
                message: Chat decision was already submitted
            minimumMessagesRequired:
              value:
                code: CHAT_MIN_MESSAGES_REQUIRED
                error: Conflict
                message: Minimum chat messages are required before approval
            mutualCancellationPending:
              value:
                code: CHAT_MUTUAL_CANCELLATION_PENDING
                error: Conflict
                message: A mutual cancellation request is pending
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, error, message]
      properties:
        code: { type: string, example: PROFILE_NOT_ACTIVE }
        error: { type: string, example: Conflict }
        message: { type: string, nullable: true }
...
    PartnerSummaryResponse:
      type: object
      required: [userId, profileId, displayName]
      properties:
        userId: { type: string, format: uuid }
        profileId: { type: string, format: uuid }
        displayName: { type: string }
...
    PhotoResponse:
      type: object
      required: [id, url, position, isPersonPhoto, isFullBody, validationStatus, moderationStatus]
      properties:
        id: { type: string, format: uuid }
        url:
          type: string
          description: Renderable photo read URL. May be a time-limited presigned URL for private object storage.
        position: { type: integer }
        isPersonPhoto: { type: boolean }
        isFullBody: { type: boolean }
        validationStatus:
          allOf: [{ $ref: '#/components/schemas/PhotoValidationStatus' }]
          description: Technical and semantic photo-validation status. Outside prod, provider none is used even if sightengine is configured, and the MVP shortcut may return VALIDATED after technical validation. In prod with provider none, technical validation alone returns PENDING and does not establish person/full-body semantics. With provider sightengine in prod, real face presence is used only as an MVP isPersonPhoto signal; at least one real face validates person-photo semantics, artificial faces do not count, and successful analysis still leaves isFullBody=false.
        moderationStatus:
          allOf: [{ $ref: '#/components/schemas/PhotoModerationStatus' }]
          description: Content moderation status. With provider none outside prod, including when sightengine is configured outside prod, APPROVED is a permissive MVP result and does not represent external safety review. With provider none in prod, uploads persist NEEDS_REVIEW. With provider sightengine in prod, Reals maps provider output to sexual explicit, sexual suggestive, violence/threat, gore and hate/extremism scores, then applies configurable policy to APPROVED, NEEDS_REVIEW or REJECTED. Automatic provider moderation does not create safety reports, blocks, penalties or bans.
    VisualProfileResponse:
      type: object
      required:
        - profileId
        - displayName
        - age
        - photos
        - myPersonalMessageSubmitted
        - partnerPersonalMessageSubmitted
        - partnerPersonalMessageRead
        - decisionRequiresPartnerPersonalMessageRead
      properties:
        profileId: { type: string, format: uuid }
        displayName: { type: string }
        age: { type: integer }
        bio: { type: string, nullable: true }
        photos:
          type: array
          items: { $ref: '#/components/schemas/PhotoResponse' }
        myPersonalMessageSubmitted:
          type: boolean
          description: True when the authenticated user has already submitted their visual personal message for this match.
        partnerPersonalMessageSubmitted:
          type: boolean
          description: True when the partner has submitted a visual personal message for this match.
        partnerPersonalMessageRead:
          type: boolean
          description: True when no partner personal message exists, or when the partner message exists and the authenticated user has fetched/read it.
        decisionRequiresPartnerPersonalMessageRead:
          type: boolean
          description: Temporarily retained for backward-compatible response shape. Always false; partner personal-message reads are no longer a visual-decision prerequisite.
        visualExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: Visual review deadline used by clients for advisory countdowns; backend remains the source of truth.
...
    MatchResponse:
      type: object
      required: [id, userAId, userBId, state, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userAId: { type: string, format: uuid }
        userBId: { type: string, format: uuid }
        state: { $ref: '#/components/schemas/MatchState' }
        connectionId: { type: string, format: uuid, nullable: true }
        visualExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChatDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision: { $ref: '#/components/schemas/ChatContinueDecision' }
    VisualDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision: { $ref: '#/components/schemas/VisualDecision' }
    PersonalMessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1, maxLength: 280, pattern: '^[^\p{Cntrl}<>]*$' }
    PartnerMessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, nullable: true }
    ChatResponse:
      type: object
      required: [id, matchId, chatType, status, startedAt, timeoutAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        matchId: { type: string, format: uuid }
        connectionId: { type: string, format: uuid, nullable: true }
        chatType: { $ref: '#/components/schemas/ChatType' }
        status: { $ref: '#/components/schemas/ChatStatus' }
        startedAt: { type: string, format: date-time }
        availableAt: { type: string, format: date-time, nullable: true }
        activatedAt: { type: string, format: date-time, nullable: true }
        conversationStartedAt: { type: string, format: date-time, nullable: true }
        timeoutAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        readOnlyUntil: { type: string, format: date-time, nullable: true }
        lastMessageAt: { type: string, format: date-time, nullable: true }
        lastMessageSenderId: { type: string, format: uuid, nullable: true }
        audioPolicy:
          allOf:
            - $ref: '#/components/schemas/ChatAudioPolicyResponse'
          nullable: true
        inactivityExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: First-chat inactivity deadline. Null for chat types where inactivity timeout does not apply.
        audioPolicy:
          allOf:
            - $ref: '#/components/schemas/ChatAudioPolicyResponse'
          nullable: true
    FirstChatResponse:
      allOf:
        - $ref: '#/components/schemas/ChatResponse'
        - type: object
          required: [partner, myDecision, partnerDecision]
          properties:
            partner: { $ref: '#/components/schemas/PartnerSummaryResponse' }
            myDecision: { $ref: '#/components/schemas/ChatParticipantDecisionStatus' }
            partnerDecision: { $ref: '#/components/schemas/ChatParticipantDecisionStatus' }
            guidance:
              allOf:
                - $ref: '#/components/schemas/FirstChatGuidanceResponse'
              nullable: true
              description: Nullable for legacy first-chat rows created before guidance existed. New first chats initialize guidance.
            audioPolicy:
              allOf:
                - $ref: '#/components/schemas/ChatAudioPolicyResponse'
              nullable: true
    ChatAudioPolicyResponse:
      type: object
      required: [enabled, maxDurationMillis, maxFileSizeBytes]
      properties:
        enabled: { type: boolean }
        unavailableReason:
          allOf:
            - $ref: '#/components/schemas/ChatAudioUnavailableReason'
          nullable: true
        enabledAt: { type: string, format: date-time, nullable: true }
        maxDurationMillis: { type: integer, format: int64, example: 60000 }
        maxFileSizeBytes: { type: integer, format: int64, example: 2097152 }
        remainingMessages:
          type: integer
          nullable: true
          description: First-chat remaining per-user audio messages; null means unlimited second-chat audio.
    FirstChatGuidanceQuestionResponse:
      type: object
      required: [id, text]
      properties:
        id: { type: string }
        text: { type: string }
    FirstChatGuidanceResponse:
      type: object
      required:
        - question
        - questionOrdinal
        - maxQuestions
        - requiredCharacters
        - canRequestNext
        - myNextRequested
        - completed
      properties:
        question:
          $ref: '#/components/schemas/FirstChatGuidanceQuestionResponse'
        questionOrdinal:
          type: integer
          minimum: 1
          maximum: 3
        maxQuestions:
          type: integer
          example: 3
        requiredCharacters:
          type: integer
          example: 40
        canRequestNext:
          type: boolean
          description: Eligibility for the authenticated user only.
        myNextRequested:
          type: boolean
          description: Whether the authenticated user has requested continuation for the current question.
        completed:
          type: boolean
          description: True when the final configured guidance question is active. The final question remains returned as `question`, and no further next-request action is available.
...
    PhotoValidationStatus:
      type: string
      enum: [PENDING, VALIDATED, FAILED]
    PhotoModerationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, NEEDS_REVIEW]
...
    MatchState:
      type: string
      enum: [CHAT_ACTIVE, VISUAL_PHASE, VISUAL_APPROVED, CHAT_REJECTED, VISUAL_REJECTED, EXPIRED]
    ChatContinueDecision:
      type: string
      enum: [APPROVED, REJECTED]
    ChatParticipantDecisionStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, ABANDONED]
    VisualDecision:
      type: string
      enum: [APPROVED, REJECTED]
    ChatType:
      type: string
      enum: [FIRST_CHAT, SECOND_CHAT]
...
    ChatAudioUnavailableReason:
      type: string
      enum: [FEATURE_DISABLED, CHAT_NOT_WRITABLE, GUIDANCE_NOT_AVAILABLE, GUIDANCE_REQUIRED, LIMIT_REACHED, WAITING_FOR_BOTH, WAITING_DELAY]
    ChatStatus:
      type: string
      enum: [AVAILABLE, ACTIVE, CANCELLED, EXPIRED, ABANDONED, CLOSED, FINISHED]
...